fix: Simplified layout.tsx by consolidating all header components into one header.tsx component
